refactor: 패키지 구조 변경 및 PaymentCardBox 컴포저블 제거
