+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardScreen.kt
@@ -11,7 +11,6 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -20,7 +19,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.data.BankRepository
-import woowacourse.payments.domain.model.BankType.Companion.toColor
+import woowacourse.payments.ui.component.CardImage
 import woowacourse.payments.ui.model.CardHolderUiModel
 import woowacourse.payments.ui.model.CardHolderUiModel.Companion.CARD_HOLDER_MAX_LENGTH
 import woowacourse.payments.ui.model.CardNumberUiModel
@@ -33,7 +32,7 @@ import woowacourse.payments.ui.newcard.components.ExpirationDateTextField
 import woowacourse.payments.ui.newcard.components.NameTextField
 import woowacourse.payments.ui.newcard.components.NewCardTopBar
 import woowacourse.payments.ui.newcard.components.PasswordField
-import woowacourse.payments.ui.newcard.components.PaymentCardBox
+import woowacourse.payments.ui.newcard.dialog.BankBottomSheet
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -89,8 +88,11 @@ fun NewCardScreen(
                     .padding(innerPadding)
                     .fillMaxSize(),
         ) {
-            PaymentCardBox(
-                background = newCardStateHolder.bank.type.toColor(),
+            CardImage(
+                bankType = newCardStateHolder.bank.type,
+                cardNumber = newCardStateHolder.cardNumber,
+                cardHolder = newCardStateHolder.cardHolder,
+                expirationDate = newCardStateHolder.expirationDateUiState.expirationDate.value,
                 modifier =
                     Modifier
                         .align(Alignment.CenterHorizontally)
