+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardStateHolder.kt
@@ -3,7 +3,6 @@ package woowacourse.payments.ui.newcard
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import kotlinx.parcelize.Parcelize
 import woowacourse.payments.domain.model.Bank
 
 class NewCardStateHolder {
