--- app/src/main/java/woowacourse/payments/ui/newcard/components/PaymentCardBox.kt
@@ -1,56 +1,0 @@
-package woowacourse.payments.ui.newcard.components
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import woowacourse.payments.R
-
-@Composable
-fun PaymentCardBox(
-    background: Int,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        contentAlignment = Alignment.CenterStart,
-        modifier =
-            modifier
-                .shadow(8.dp)
-                .size(width = 208.dp, height = 124.dp)
-                .background(
-                    color = Color(background),
-                    shape = RoundedCornerShape(5.dp),
-                ),
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .padding(start = 14.dp, bottom = 10.dp)
-                    .size(width = 40.dp, height = 26.dp)
-                    .background(
-                        color = colorResource(R.color.card_chip_gold),
-                        shape = RoundedCornerShape(4.dp),
-                    ),
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun PaymentCardBoxPreview() {
-    PaymentCardBox(
-        background = R.color.card_background,
-        modifier =
-            Modifier
-                .padding(top = 14.dp),
-    )
-}
